Started genericizing API result types for better exception handling
We now have two packages, rapture.io and rapture.iox.  The former works
exactly as before, and throws exceptions when something goes wrong.

The new rapture.iox package, however, has exactly the same API, except
that instead of returning type T for some T, it returns
Either[Exception, T], capturing the result in the Right branch of
Either, or the exception in the Left branch.